Added the beginning of PID control

+++ include/kovan/motors.hpp
@@ -47,9 +47,9 @@ public:
 	 */
 	Motor(const port_t& port) throw();
 	
-	void moveAtVelocity(const int& velocity);
-	void moveToPosition(const int& speed, const int& goalPos);
-	void moveRelativePosition(const int& speed, const int& deltaPos);
+	void moveAtVelocity(const short& velocity);
+	void moveToPosition(const short& speed, const short& goalPos);
+	void moveRelativePosition(const short& speed, const short& deltaPos);
 	
 	void setPidGains(const short& p, const short& i, const short& d, const short& pd, const short& id, const short& dd);
 	void pidGains(short& p, short& i, short& d, short& pd, short& id, short& dd);
